requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
Werkzeug==3.1.3
redis==6.0.0
structlog==25.3.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
aiosqlite==0.21.0
requests==2.32.3
icalendar==6.1.3
reportlab==4.4.1
PyPDF2==3.0.1
fillpdf==0.7.3
